# infrastructure/gateway/src/main/resources/openapi/documentation.yaml
openapi: 3.0.3
info:
  title: Meldestelle API
  description: |
    Self-Contained Systems API Gateway for Austrian Equestrian Federation.
    This API provides access to various bounded contexts including authentication,
    master data management, horse registry, and event management.

    ## Rate Limiting
    This API implements rate limiting to ensure fair usage and system stability.

    ### Global Rate Limits
    - Default limit: 100 requests per minute per IP address
    - When the limit is exceeded, the API will respond with a 429 Too Many Requests status code

    ### User-Based Rate Limits
    Different rate limits apply based on user authentication status:
    - Anonymous users: 50 requests per minute
    - Authenticated users: 200 requests per minute
    - Admin users: 500 requests per minute

    ### Endpoint-Specific Rate Limits
    Some endpoints have specific rate limits:
    - Authentication endpoints: 20 requests per minute
    - Event management endpoints: 200 requests per minute

    ### Rate Limit Headers
    The API includes rate limit information in response headers:
    - `X-RateLimit-Enabled`: Indicates if rate limiting is enabled
    - `X-RateLimit-Limit`: The rate limit ceiling for the given endpoint
    - `X-Correlation-ID`: A unique identifier for request tracking

    ### Best Practices
    - Implement exponential backoff when receiving 429 responses
    - Cache responses when appropriate to reduce API calls
...

    ## Enhanced Logging
...

...
  contact:
...
servers:
...
paths:
...
  /api/horses:
    get:
      tags:
        - Horse Registry
      summary: Get All Horses
      description: Returns a list of all horses
      operationId: getAllHorses
      security:
        - bearerAuth: []
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Filter to only return active horses
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
          description: Maximum number of horses to return
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter horses by owner ID
        - name: geschlecht
          in: query
          required: false
          schema:
            type: string
            enum: [STALLION, MARE, GELDING]
          description: Filter horses by gender
        - name: rasse
          in: query
          required: false
          schema:
            type: string
          description: Filter horses by breed
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term to filter horses by name
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorsesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/search/lebensnummer/{nummer}:
    get:
      tags:
        - Horse Registry
      summary: Find Horse by Life Number
      description: Returns a horse by its life number
      operationId: getHorseByLifeNumber
      security:
        - bearerAuth: []
      parameters:
        - name: nummer
          in: path
          required: true
          schema:
            type: string
          description: Life number of the horse
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/search/chip/{nummer}:
    get:
      tags:
        - Horse Registry
      summary: Find Horse by Chip Number
      description: Returns a horse by its chip number
      operationId: getHorseByChipNumber
      security:
        - bearerAuth: []
      parameters:
        - name: nummer
          in: path
          required: true
          schema:
            type: string
          description: Chip number of the horse
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/oeps-registered:
    get:
      tags:
        - Horse Registry
      summary: Get OEPS Registered Horses
      description: Returns a list of horses registered with the Austrian Equestrian Federation (OEPS)
      operationId: getOepsRegisteredHorses
      security:
        - bearerAuth: []
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Filter to only return active horses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorsesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/fei-registered:
    get:
      tags:
        - Horse Registry
      summary: Get FEI Registered Horses
      description: Returns a list of horses registered with the International Federation for Equestrian Sports (FEI)
      operationId: getFeiRegisteredHorses
      security:
        - bearerAuth: []
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Filter to only return active horses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorsesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/stats:
    get:
      tags:
        - Horse Registry
      summary: Get Horse Statistics
      description: Returns statistics about horses in the registry
      operationId: getHorseStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseStatsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Horse Registry
      summary: Register Horse
      description: Registers a new horse
      operationId: registerHorse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterHorseRequest'
      responses:
        '201':
          description: Horse successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseResponse'
        '400':
          description: Invalid horse data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/{id}:
    get:
      tags:
        - Horse Registry
      summary: Get Horse by ID
      description: Returns a horse by its ID
      operationId: getHorseById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Horse Registry
      summary: Update Horse
      description: Updates an existing horse
      operationId: updateHorse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHorseRequest'
      responses:
        '200':
          description: Horse successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorseResponse'
        '400':
          description: Invalid horse data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Horse Registry
      summary: Delete Horse
      description: Deletes a horse
      operationId: deleteHorse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Force delete even if the horse has dependencies
      responses:
        '200':
          description: Horse successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/{id}/soft-delete:
    post:
      tags:
        - Horse Registry
      summary: Soft Delete Horse
      description: Marks a horse as inactive instead of permanently deleting it
      operationId: softDeleteHorse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Horse successfully marked as inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Horse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/horses/batch-delete:
    post:
      tags:
        - Horse Registry
      summary: Batch Delete Horses
      description: Deletes multiple horses in a single operation
      operationId: batchDeleteHorses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDeleteRequest'
      responses:
        '200':
          description: Horses successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeleteResponse'
        '206':
          description: Partial content - some horses could not be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeleteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Event Management Context
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - error
      properties:
        success:
          type: boolean
          default: false
          description: Indicates that the operation failed
        message:
          type: string
          description: A message describing the error
        error:
          type: string
          description: The error code or type
...
    HorsesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Horse'
    HorseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Horse'
    RegisterHorseRequest:
      type: object
      required:
        - name
        - birthDate
        - breed
        - color
        - gender
      properties:
        name:
          type: string
          example: Maestoso
        birthDate:
          type: string
          format: date
          example: 2015-05-15
        breed:
          type: string
          example: Lipizzaner
        color:
          type: string
          example: White
        gender:
          type: string
          enum: [STALLION, MARE, GELDING]
          example: STALLION
        fatherId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        motherId:
          type: string
          format: uuid
          example: 223e4567-e89b-12d3-a456-426614174000
        ownerId:
          type: string
          format: uuid
          example: 323e4567-e89b-12d3-a456-426614174000
        registrationNumber:
          type: string
          example: AT-LIP-2015-123
...
    HorseStatsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/HorseStats'
...
    BatchDeleteResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BatchDeleteResult'
...